+++ src/formats.ts
@@ -32,12 +32,12 @@ export const defaultColorMatches = (
     coverage >= style.stage1
       ? 76
       : coverage >= style.stage2
-        ? Math.floor(
+      ? Math.floor(
           ((style.stage1 - coverage) / (style.stage1 - style.stage2)) * 10
         ) *
-        16 +
+          16 +
         76
-        : 225 + Math.floor(coverage / (style.stage2 / 11));
+      : 225 + Math.floor(coverage / (style.stage2 / 11));
   const result = gradient.toString(16);
   return (result.length === 1 ? "0" : "") + result + "1";
 };
@@ -75,11 +75,11 @@ const FormatsObj: FormatObj = {
     }
   },
 
-  listFormats: function () {
+  listFormats: function() {
     return Object.keys(this.formats);
   },
 
-  getFormat: function (format: string) {
+  getFormat: function(format: string) {
     return this.formats[format];
   }
 };